Create README.md
Hi everyone!  Looking forward to being a part of this big creative world of developing the future with likeminded software developers.  I hope to soon be a part of some interesting open source projects for the good of the community.  If you want to connect, I am also on Discord:  TheNDCoder #9375.  :-)